Implemented filtering by geo distance

+++ src/Internal/Geo.php
@@ -12,9 +12,6 @@ class Geo
     public static function geoDistance(float $latA, float $lngA, float $latB, float $lngB): float
     {
         // Use Haversine here, it's faster
-        //dump((new Coordinate($latA, $lngA))->getDistance(new Coordinate($latB, $lngB), new Haversine()));
-        // dump((int) round((new Coordinate($latA, $lngA))->getDistance(new Coordinate($latB, $lngB), new Haversine())));
-
-        return (float) (new Coordinate($latA, $lngA))->getDistance(new Coordinate($latB, $lngB), new Haversine());
+        return (new Coordinate($latA, $lngA))->getDistance(new Coordinate($latB, $lngB), new Haversine());
     }
 }
